Fixing destination modification assumptions
While this is breaking for applications that depends on the ldp parameter, there should not be any assumptions that it should be included, as it causes implementations with simplesamlserver and google to fail. If the destination address requires the parameter to be set, it should be manually set in the applications config options.