Fixing destination modification assumptions (#22)
* Fixing destination modification assumptions

While this is breaking for applications that depends on the ldp parameter, there should not be any assumptions that it should be included, as it causes implementations with simplesamlserver and google to fail. If the destination address requires the parameter to be set, it should be manually set in the applications config options.

* Adding new query_params to config template

* Adding query parameter configuration

If the `query_params` config option is not set for that sp, it will default to the previous functionality. This will make sure any applications depending on the current functionality will not break.

* Removing unnecessary storage class use

* Add query params to logout destination